grid-demo-domain/src/main/java/no/jaf/demo/grid/domain/Person.java
package no.jaf.demo.grid.domain;

import java.io.Serializable;

public class Person implements Serializable {
    private SocialSecurityNumber socialSecurityNumber;
    private String name;
    private Address address;

    public SocialSecurityNumber getSocialSecurityNumber() {
        return socialSecurityNumber;
    }

    public void setSocialSecurityNumber(SocialSecurityNumber socialSecurityNumber) {
        this.socialSecurityNumber = socialSecurityNumber;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public Address getAddress() {
        return address;
    }

    public void setAddress(Address address) {
        this.address = address;
    }
}
